 requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
aiogram==3.20.0
pydantic==2.11.3